set default to 64 bit integers

+++ include/typedefs.hpp
@@ -4,7 +4,7 @@
 #include <cstdint>
 #include "../include/types/int40.hpp"
 
-typedef int40 c_size_t;
+typedef int64_t c_size_t;
 typedef uint32_t uc_size_t;
 typedef int64_t len_t;
 typedef uint8_t char_t;
